feat(bookmark): migrate functionality from UserService to BookmarksService
