+++ itachallenge-user/src/main/java/com/itachallenge/userinteraction/service/bookmark/BookmarkService.java
@@ -7,4 +7,6 @@ import java.util.UUID;
 
 public interface BookmarkService {
     Mono<Set<UUID>> getUserBookmarks(String userId);
+    Mono<Boolean> addChallengeToBookmarks(String userId, String challengeId);
+    Mono<Boolean> deleteChallengeFromBookmarks(String userId, String challengeId);
 }
