+++ itachallenge-user/src/main/java/com/itachallenge/userinteraction/service/bookmark/BookmarkServiceImpl.java
@@ -38,6 +38,57 @@ public class BookmarkServiceImpl implements BookmarkService {
                 );
     }
 
+    @Override
+    public Mono<Boolean> addChallengeToBookmarks(String userId, String challengeId) {
+        return Mono.zip(parseAndValidateUUID(userId), parseAndValidateUUID(challengeId))
+                .flatMap(uuidTuple -> {
+                    UUID userUuid = uuidTuple.getT1();
+                    UUID challengeUuid = uuidTuple.getT2();
+
+                    return userRepository.findById(userUuid)
+                            .switchIfEmpty(Mono.error(new NotFoundException("User not found")))
+                            .flatMap(user -> addToBookmarks(userUuid, challengeUuid));
+                });
+    }
+
+    @Override
+    public Mono<Boolean> deleteChallengeFromBookmarks(String userId, String challengeId) {
+        return Mono.zip(parseAndValidateUUID(userId), parseAndValidateUUID(challengeId))
+                .flatMap(uuidTuple -> {
+                    UUID userUuid = uuidTuple.getT1();
+                    UUID challengeUuid = uuidTuple.getT2();
+
+                    return userRepository.findById(userUuid)
+                            .switchIfEmpty(Mono.error(new NotFoundException("User not found")))
+                            .flatMap(user -> deleteFromBookmarks(userUuid, challengeUuid));
+                });
+    }
+
+    private Mono<Boolean> addToBookmarks(UUID userUuid, UUID challengeUuid) {
+        return bookmarkRepository.existsByUserIdAndChallengeId(userUuid, challengeUuid)
+                .flatMap(exists -> {
+                    if (exists.booleanValue())
+                        return Mono.just(false);
+
+                    BookmarkDocument bookmark = new BookmarkDocument();
+                    bookmark.setUuid(UUID.randomUUID());
+                    bookmark.setUserId(userUuid);
+                    bookmark.setChallengeId(challengeUuid);
+
+                    return bookmarkRepository.save(bookmark)
+                            .thenReturn(true);
+                });
+    }
+
+    private Mono<Boolean> deleteFromBookmarks(UUID userId, UUID challengeUuid) {
+        return bookmarkRepository.findByUserIdAndChallengeId(userId, challengeUuid)
+                .flatMap(bookmark ->
+                        bookmarkRepository.delete(bookmark)
+                                .then(Mono.just(true))
+                )
+                .switchIfEmpty(Mono.just(false));
+    }
+
     private Mono<UUID> parseAndValidateUUID(String id) {
         if (id == null || id.isEmpty()) {
             return Mono.error(new BadUUIDException("Invalid ID format"));
